Unit test build.gradle generation into new file
* Moved the build.gradle template and code to generate it for test into it's own file.
